+++ reverse_proxy.go
@@ -39,28 +39,21 @@ func newRedisPool(server, password string) *redis.Pool {
 	}
 }
 
-func wmsReverseProxy(redisHost string, redisPort string, subgridWmsPort string, flowWmsPort string, useCache bool, singleServer string) *httputil.ReverseProxy {
+func wmsReverseProxy(redisHost string, redisPort string, useCache bool, singleServer string) *httputil.ReverseProxy {
 	// normally use redis to find subgrid_id and ip belonging to session
 	// if singleServer is defined, subgrid_id = singleServer and redirect to localhost
 	var subgridID string
-	var wmsIP string
 
 	sessionKeyCache := make(map[string]string)
 
 	redisAddress := strings.Join([]string{redisHost, redisPort}, ":")
 	pool = newRedisPool(redisAddress, "")
 
-	if singleServer != "" {
-		subgridID = singleServer
-		wmsIP = "127.0.0.1"
-		log.Println("- INFO - single server mode, subgrid_id:", subgridID)
-	}
-
 	director := func(req *http.Request) {
 		var startTime = time.Now()
 		log.Println("- INFO - start resolving wms address")
 		// 1) get the session key
-		var sessionKey, wmsPort string
+		var sessionKey, addressKey, dataAddress string
 
 		sessionCookie, err := req.Cookie("sessionid")
 		remoteAddr := strings.Split(req.RemoteAddr, ":")[0] // skip the port
@@ -100,17 +93,12 @@ func wmsReverseProxy(redisHost string, redisPort string, subgridWmsPort string, 
 				return
 			}
 			log.Println("- INFO - got subgrid id:", subgridID)
-
-			// 3) get the wms server ip
-			wmsIP, err = redis.String(conn.Do("HGET", "subgrid_id_to_ip", subgridID))
-			if err != nil {
-				log.Println("- ERROR - unable to get wms ip:", err)
-				req.URL = nil
-				return
-			}
+		} else {
+			subgridID = singleServer
+			log.Println("- INFO - single server mode, subgrid_id:", subgridID)
 		}
 
-		// 4) get the loaded_model_type
+		// 3) get the loaded_model_type
 		loadedModelType, err := redis.String(conn.Do("GET", subgridID+":loaded_model_type"))
 		if err != nil {
 			// assume loaded_model_type is 3di for backwards compatibility
@@ -119,25 +107,46 @@ func wmsReverseProxy(redisHost string, redisPort string, subgridWmsPort string, 
 		}
 		log.Println("- INFO - using loaded model type:", loadedModelType)
 
-		// 5) determine which wms port to use based on the loaded model type
+		// 4) determine which address key to use based on the loaded model type
 		switch loadedModelType {
 		case "3di":
-			wmsPort = subgridWmsPort
+			addressKey = "subgrid_id_to_wms_address"
 		case "3di-urban":
-			wmsPort = flowWmsPort
-		// assume loaded_model_type is 3di for backwards compatibility
+			addressKey = "subgrid_id_to_flow_address"
 		default:
 			log.Println("- ERROR - unsupported loaded model type:", loadedModelType)
 			req.URL = nil
 			return
 		}
 
-		// use the full wms address to redirect this request to
-		wmsAddress := strings.Join([]string{wmsIP, wmsPort}, ":")
+		// 5) get the data address
+		if singleServer == "" {
+			dataAddress, err = redis.String(conn.Do("HGET", addressKey, subgridID))
+			if err != nil {
+				log.Println("- ERROR - unable to get data address:", err)
+				req.URL = nil
+				return
+			}
+		} else {
+			switch loadedModelType {
+			case "3di":
+				// 5000 is the default wms port
+				dataAddress = "127.0.0.1:5000"
+			case "3di-urban":
+				// 6000 is the default flow port
+				dataAddress = "127.0.0.1:6000"
+			default:
+				log.Println("- ERROR - unsupported loaded model type:", loadedModelType)
+				req.URL = nil
+				return
+			}
+		}
+
+		// redirect this request to the resolved data address
 		var endTime = time.Now()
-		log.Println("- INFO - resolved wms address", wmsAddress, "in", endTime.Sub(startTime))
+		log.Println("- INFO - resolved data address", dataAddress, "in", endTime.Sub(startTime))
 		req.URL.Scheme = "http"
-		req.URL.Host = wmsAddress
+		req.URL.Host = dataAddress
 	}
 	return &httputil.ReverseProxy{Director: director}
 }