+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of wmsrp
 0.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Resolve full address (<host or ip>:<port>) to redirect to instead of ip only.
 
 
 0.1.0 (2015-08-03)